+++ Sources/KeyboardKit/Feedback/FeedbackConfiguration.swift
@@ -13,6 +13,9 @@ import Foundation
  This class can be used to specify what kind of feedback the
  current keyboard should give to the user.
  
+ Use ``audioConfiguration`` and ``hapticConfiguration`` when
+ modifying the feedback behavior.
+ 
  KeyboardKit automatically creates an instance of this class
  and binds it to ``KeyboardInputViewController/state``.
  */
@@ -24,48 +27,37 @@ public class FeedbackConfiguration: ObservableObject {
      - Parameters:
        - audioConfiguration: The configuration to use for audio feedback, by deafult `.enabled`.
        - hapticConfiguration: The configuration to use for haptic feedback, by deafult `.minimal`.
-       - enabledAudioConfiguration: The configuration to use to enable audio feedback, by default `.enabled`.
-       - enabledHapticConfiguration: The configuration to use to enable haptic feedback, by default `.enabled`.
-       - disabledAudioConfiguration: The configuration to use to disable audio feedback, by default `.standard`.
-       - disabledHapticConfiguration: The configuration to use to disable haptic feedback, by default `.minimal`.
      */
     public init(
-        audioConfiguration: AudioFeedback.Configuration = .enabled,
-        hapticConfiguration: HapticFeedback.Configuration = .minimal,
-        enabledAudioConfiguration: AudioFeedback.Configuration = .enabled,
-        enabledHapticConfiguration: HapticFeedback.Configuration = .enabled,
-        disabledAudioConfiguration: AudioFeedback.Configuration = .enabled,
-        disabledHapticConfiguration: HapticFeedback.Configuration = .minimal
+        audio: AudioFeedback.Configuration = .enabled,
+        haptic: HapticFeedback.Configuration = .minimal
     ) {
-        self.audioConfiguration = audioConfiguration
-        self.hapticConfiguration = hapticConfiguration
-        self.enabledAudioConfiguration = enabledAudioConfiguration
-        self.enabledHapticConfiguration = enabledHapticConfiguration
-        self.disabledAudioConfiguration = disabledAudioConfiguration
-        self.disabledHapticConfiguration = disabledHapticConfiguration
+        self.audio = audio
+        self.haptic = haptic
+        enabledAudio = audio == .disabled ? .enabled : audio
+        enabledHaptic = haptic == .disabled ? .enabled : haptic
     }
     
-    
-    /// The configuration to use to enable audio feedback.
-    public var enabledAudioConfiguration: AudioFeedback.Configuration
-    
-    /// The configuration to use to enable haptic feedback.
-    public var enabledHapticConfiguration: HapticFeedback.Configuration
-    
-    /// The configuration to use to disable audio feedback.
-    public var disabledAudioConfiguration: AudioFeedback.Configuration
-    
-    /// The configuration to use to disable haptic feedback.
-    public var disabledHapticConfiguration: HapticFeedback.Configuration
-    
-
     /// The configuration to use for audio feedback.
     @Published
-    public var audioConfiguration: AudioFeedback.Configuration
+    public var audio: AudioFeedback.Configuration {
+        didSet {
+            guard audio != .disabled else { return }
+            enabledAudio = audio
+        }
+    }
     
     /// The configuration to use for haptic feedback.
     @Published
-    public var hapticConfiguration: HapticFeedback.Configuration
+    public var haptic: HapticFeedback.Configuration {
+        didSet {
+            guard haptic != .disabled else { return }
+            enabledHaptic = haptic
+        }
+    }
+    
+    private var enabledAudio: AudioFeedback.Configuration
+    private var enabledHaptic: HapticFeedback.Configuration
 }
 
 public extension FeedbackConfiguration {
@@ -78,51 +70,23 @@ public extension FeedbackConfiguration {
 
     /// Get or set whether or not audio feedback is enabled.
     var isAudioFeedbackEnabled: Bool {
-        get { audioConfiguration != .disabled }
-        set { audioConfiguration = newValue ? enabledAudioConfiguration : disabledAudioConfiguration }
+        get { audioConfiguration == enabledAudio }
+        set { audioConfiguration = newValue ? enabledAudio : .disabled }
     }
 
     /// Get or set whether or not haptic feedback is enabled.
     var isHapticFeedbackEnabled: Bool {
-        get { hapticConfiguration != .disabled }
-        set { hapticConfiguration = newValue ? enabledHapticConfiguration : disabledHapticConfiguration }
-    }
-
-    /// Disable audio feedback.
-    func disableAudioFeedback() {
-        isAudioFeedbackEnabled = false
-    }
-
-    /// Disable haptic feedback.
-    func disableHapticFeedback() {
-        isHapticFeedbackEnabled = false
-    }
-
-    /// Enable audio feedback.
-    func enableAudioFeedback() {
-        isAudioFeedbackEnabled = true
-    }
-
-    /// Enable haptic feedback.
-    func enableHapticFeedback() {
-        isHapticFeedbackEnabled = true
+        get { hapticConfiguration == enabledHaptic }
+        set { hapticConfiguration = newValue ? enabledHaptic : .disabled }
     }
 
     /// Toggle audio feedback between enabled and disabled.
-    func toggleAudioFeedback() {
-        if isAudioFeedbackEnabled {
-            disableAudioFeedback()
-        } else {
-            enableAudioFeedback()
-        }
+    func toggleIsAudioFeedbackEnabled() {
+        isAudioFeedbackEnabled.toggle()
     }
 
     /// Toggle haptic feedback between enabled and disabled.
-    func toggleHapticFeedback() {
-        if isHapticFeedbackEnabled {
-            disableHapticFeedback()
-        } else {
-            enableHapticFeedback()
-        }
+    func toggleIsHapticFeedbackEnabled() {
+        isHapticFeedbackEnabled.toggle()
     }
 }